Finally make the request framework something I am proud of
Get rid of all that old crap. Enter ApiUtility

ApiUtility is a singleton that holds parsed matches. It is always ready to offer up a Match whenever called and will never run out of matches to give. It goes five minutes into the past to get more matches each time the stack of matches is about to be exhausted (2 matches left)

TODO: animations only work on first load, and the background doesnt change either. But this doesnt have to do with the new ApiUtility