URF is over. No hope of pulling live match data
So do this instead. This is (IMO) pretty cool. Take the last known timestamp for URF matches, and pick one of 3500 timestamps before it at random!

Also, rework the way champion backgrounds are processed. determine the MVP and start downloading its skin right away, since it takes the longest to download.